change `sys_cycle_count` to return `u64`
Changing `sys_cycle_count` resulted in various changes to the trait as
well as several other type changes throughout the system including
`SyscallContext`. I've addressed various other type errors as well as
`TODO(breaking change): update to `u64`` comments. Tests have also been
updated to reflect this type change.

+++ risc0/zkvm/src/guest/env.rs
@@ -539,7 +539,7 @@ pub fn commit_slice<T: Pod>(slice: &[T]) {
 /// began.
 ///
 /// WARNING: The cycle count is provided by the host and is not checked by the zkVM circuit.
-pub fn cycle_count() -> usize {
+pub fn cycle_count() -> u64 {
     sys_cycle_count()
 }
